1345 - changed names from AiTextAnalysis to AiText
