+++ server/ee/apps/worker-app/build.gradle.kts
@@ -50,7 +50,7 @@ dependencies {
     implementation(project(":server:libs:modules:components:accelo"))
     implementation(project(":server:libs:modules:components:active-campaign"))
     implementation(project(":server:libs:modules:components:affinity"))
-    implementation(project(":server:libs:modules:components:ai:ai-text-analysis"))
+    implementation(project(":server:libs:modules:components:ai:ai-text"))
     implementation(project(":server:libs:modules:components:ai:llm:amazon-bedrock"))
     implementation(project(":server:libs:modules:components:ai:llm:anthropic"))
     implementation(project(":server:libs:modules:components:ai:llm:azure-openai"))
