--- server/libs/modules/components/ai/ai-text-analysis/src/main/java/com/bytechef/component/ai/text/analysis/action/ClassifyTextAction.java
@@ -1,111 +1,0 @@
-/*
- * Copyright 2023-present ByteChef Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.bytechef.component.ai.text.analysis.action;
-
-import static com.bytechef.component.ai.llm.constant.LLMConstants.MAX_TOKENS_PROPERTY;
-import static com.bytechef.component.ai.llm.constant.LLMConstants.MODEL;
-import static com.bytechef.component.ai.llm.constant.LLMConstants.TEMPERATURE_PROPERTY;
-import static com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants.CATEGORIES;
-import static com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants.EXAMPLES;
-import static com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants.MODEL_NO_OPTIONS_PROPERTY;
-import static com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants.MODEL_OPTIONS_PROPERTY;
-import static com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants.MODEL_PROVIDER_PROPERTY;
-import static com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants.MODEL_URL_PROPERTY;
-import static com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants.TEXT;
-import static com.bytechef.component.definition.ComponentDsl.action;
-import static com.bytechef.component.definition.ComponentDsl.array;
-import static com.bytechef.component.definition.ComponentDsl.object;
-import static com.bytechef.component.definition.ComponentDsl.string;
-
-import com.bytechef.component.ai.text.analysis.action.definition.AiTextAnalysisActionDefinition;
-import com.bytechef.component.ai.text.analysis.constant.AiTextAnalysisConstants;
-import com.bytechef.component.definition.Parameters;
-import com.bytechef.config.ApplicationProperties;
-import com.bytechef.platform.component.definition.ParametersFactory;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-/**
- * @author Marko Kriskovic
- */
-public class ClassifyTextAction implements AiTextAnalysisAction {
-
-    public final AiTextAnalysisActionDefinition actionDefinition;
-
-    public ClassifyTextAction(ApplicationProperties.Ai.Component component) {
-        this.actionDefinition = getActionDefinition(component);
-    }
-
-    private AiTextAnalysisActionDefinition getActionDefinition(ApplicationProperties.Ai.Component component) {
-        return new AiTextAnalysisActionDefinition(
-            action(AiTextAnalysisConstants.CLASSIFY_TEXT)
-                .title("Classify Text")
-                .description("AI reads, analyzes and classifies your text into one of defined categories.")
-                .properties(
-                    MODEL_PROVIDER_PROPERTY,
-                    MODEL_OPTIONS_PROPERTY,
-                    MODEL_NO_OPTIONS_PROPERTY,
-                    MODEL_URL_PROPERTY,
-                    string(TEXT)
-                        .label("Text")
-                        .description("The text that is to be classified.")
-                        .required(true),
-                    array(CATEGORIES)
-                        .label("Categories")
-                        .description("A list of categories that the model can choose from.")
-                        .items(string())
-                        .required(true),
-                    object(EXAMPLES)
-                        .label("Examples")
-                        .description(
-                            "You can classify a few samples, to guide your model on how to classify the real data.")
-                        .additionalProperties(string()),
-                    MAX_TOKENS_PROPERTY,
-                    TEMPERATURE_PROPERTY)
-                .output(),
-            component, this);
-    }
-
-    public Parameters createParameters(Parameters inputParameters) {
-        Map<String, Object> modelInputParametersMap = new HashMap<>();
-
-        String systemPrompt =
-            "You will receive a list of categories, text and examples. You will choose which of the given categories fits the given text the most. Your response will only be the chosen category.";
-
-        String userBuilder = "List of categories: " + inputParameters.getList(CATEGORIES) + "\nText: " +
-            inputParameters.getString(TEXT);
-
-        Map<String, ?> exampleMap = inputParameters.getMap(EXAMPLES, String.class, Map.of());
-
-        if (!exampleMap.isEmpty()) {
-            userBuilder = userBuilder + "\nExamples: " +
-                exampleMap.entrySet()
-                    .stream()
-                    .map(Map.Entry::toString)
-                    .collect(Collectors.joining(";"));
-        }
-
-        modelInputParametersMap.put(
-            "messages",
-            List.of(Map.of("content", systemPrompt, "role", "system"), Map.of("content", userBuilder, "role", "user")));
-        modelInputParametersMap.put("model", inputParameters.getString(MODEL));
-
-        return ParametersFactory.createParameters(modelInputParametersMap);
-    }
-}
