+++ server/libs/modules/components/ai/ai-text/build.gradle.kts
